clean.sh: add a warning when the registry is not empty

+++ clean.sh
@@ -1,3 +1,13 @@
 #!/bin/bash
+if [ -f "/etc/goalias/registry.txt" ] && [ $(wc -l < /etc/goalias/registry.txt) != 0 ]; then
+	read -p "Registry is not empty; continue? (Y/N)" -n 1 -r
+	echo
+	if ! [[ $REPLY =~ ^[Yy]$ ]]
+	then
+		exit 0
+	fi
+fi
+
 sudo rm -rf /tmp/goalias
 sudo rm -rf /etc/goalias
+echo "cleaned"
